Relax version restriction on unmaintained plist cookbook
This feels like an anti-pattern because upstream cookbook could break us otherwise!

However, carsomyr/chef-plist cookbook is not well maintained!
Since there has been no movement on carsomyr/chef-plist#2 we cannot lock to a newer version via Supermarket!

If we try to lock to pre-released version number, we cause librarian-chef & berkshelf dependency resolution errors.
Yet, we could try to lock to the pre-release version on GitHub with the committed bugfixes.
This requires use of GitHub as cookbook source in `Cheffile` or `Berksfile`, which is "hidden knowledge" until a
user of this cookbook runs into errors with `plist_file`!

So, unfortunately we must relax this restriction for now and rely on the "hidden knowledge" being the workaround :-(